src/app/shared/components/popularTags/store/reducers.ts
import {routerNavigationAction} from '@ngrx/router-store'
import {createFeature, createReducer, on} from '@ngrx/store'
import {popularTagsActions} from 'src/app/shared/components/popularTags/store/actions'
import {PopularTagsStateInterface} from 'src/app/shared/components/popularTags/types/popularTagsState.interface'

export const initialState: PopularTagsStateInterface = {
  isLoading: false,
  error: null,
  data: null,
}

export const popularTagsFeature = createFeature({
  name: 'popularTags',
  reducer: createReducer(
    initialState,
    on(popularTagsActions.getPopularTags, (state) => ({
      ...state,
      isLoading: true,
    })),
    on(popularTagsActions.getPopularTagsSuccess, (state, action) => ({
      ...state,
      isLoading: false,
      data: action.popularTags,
    })),
    on(popularTagsActions.getPopularTagsFailure, (state) => ({
      ...state,
      isLoading: false,
    }))
  ),
})

export const {
  name: popularTagsFeatureKey,
  reducer: popularTagsReducer,
  selectIsLoading,
  selectError,
  selectData: selectPopularTagsData,
} = popularTagsFeature
